Move SignIn folder and added "event.preventDefault()" to login part
